# blog-service/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "博客系统",
    "contact": {},
    "license": {},
    "version": "1.0"
  },
  "paths": {
    "/api/v1/articles": {
      "get": {
        "produces": [
          "application/json"
        ],
        "summary": "获取多个文章",
        "parameters": [
          {
            "type": "string",
            "description": "文章名称",
            "name": "name",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "标签ID",
            "name": "tag_id",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "状态",
            "name": "state",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "每页数量",
            "name": "page_size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "$ref": "#/definitions/model.ArticleSwagger"
            }
          },
          "400": {
            "description": "请求错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          },
          "500": {
            "description": "内部错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          }
        }
      },
      "post": {
        "produces": [
          "application/json"
        ],
        "summary": "创建文章",
        "parameters": [
          {
            "description": "标签ID",
            "name": "tag_id",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "文章标题",
            "name": "title",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "文章简述",
            "name": "desc",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "封面图片地址",
            "name": "cover_image_url",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "文章内容",
            "name": "content",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "创建者",
            "name": "created_by",
            "in": "body",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "状态",
            "name": "state",
            "in": "body",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "$ref": "#/definitions/model.Article"
            }
          },
          "400": {
            "description": "请求错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          },
          "500": {
            "description": "内部错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          }
        }
      }
    },
    "/api/v1/articles/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "summary": "获取单个文章",
        "parameters": [
          {
            "type": "integer",
            "description": "文章ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "$ref": "#/definitions/model.Article"
            }
          },
          "400": {
            "description": "请求错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          },
          "500": {
            "description": "内部错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          }
        }
      },
      "put": {
        "produces": [
          "application/json"
        ],
        "summary": "更新文章",
        "parameters": [
          {
            "description": "标签ID",
            "name": "tag_id",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "文章标题",
            "name": "title",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "文章简述",
            "name": "desc",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "封面图片地址",
            "name": "cover_image_url",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "文章内容",
            "name": "content",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "修改者",
            "name": "modified_by",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "$ref": "#/definitions/model.Article"
            }
          },
          "400": {
            "description": "请求错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          },
          "500": {
            "description": "内部错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          }
        }
      },
      "delete": {
        "produces": [
          "application/json"
        ],
        "summary": "删除文章",
        "parameters": [
          {
            "type": "integer",
            "description": "文章ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "请求错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          },
          "500": {
            "description": "内部错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          }
        }
      }
    },
    "/api/v1/tags": {
      "get": {
        "produces": [
          "application/json"
        ],
        "summary": "获取多个标签",
        "parameters": [
          {
            "maxLength": 100,
            "type": "string",
            "description": "标签名称",
            "name": "name",
            "in": "query"
          },
          {
            "enum": [
              0,
              1
            ],
            "type": "integer",
            "default": 1,
            "description": "标签名称",
            "name": "state",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "每页个数",
            "name": "page_size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "$ref": "#/definitions/model.TagSwagger"
            }
          },
          "400": {
            "description": "请求错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          },
          "500": {
            "description": "内部错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          }
        }
      },
      "post": {
        "produces": [
          "application/json"
        ],
        "summary": "新增标签",
        "parameters": [
          {
            "maxLength": 100,
            "minLength": 3,
            "description": "标签名称",
            "name": "name",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "enum": [
              0,
              1
            ],
            "default": 1,
            "description": "状态",
            "name": "state",
            "in": "body",
            "schema": {
              "type": "integer"
            }
          },
          {
            "maxLength": 100,
            "minLength": 3,
            "description": "创建者",
            "name": "created_by",
            "in": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "$ref": "#/definitions/model.Tag"
            }
          },
          "400": {
            "description": "请求错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          },
          "500": {
            "description": "内部错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          }
        }
      }
    },
    "/api/v1/tags/{id}": {
      "put": {
        "produces": [
          "application/json"
        ],
        "summary": "更新标签",
        "parameters": [
          {
            "type": "integer",
            "description": "标签ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "maxLength": 100,
            "minLength": 3,
            "description": "标签名称",
            "name": "name",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "enum": [
              0,
              1
            ],
            "default": 1,
            "description": "状态",
            "name": "state",
            "in": "body",
            "schema": {
              "type": "integer"
            }
          },
          {
            "maxLength": 100,
            "minLength": 3,
            "description": "修改者",
            "name": "modified_by",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/model.Tag"
              }
            }
          },
          "400": {
            "description": "请求错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          },
          "500": {
            "description": "内部错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          }
        }
      },
      "delete": {
        "produces": [
          "application/json"
        ],
        "summary": "删除标签",
        "parameters": [
          {
            "type": "integer",
            "description": "标签ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "请求错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          },
          "500": {
            "description": "内部错误",
            "schema": {
              "$ref": "#/definitions/errcode.Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "app.Pager": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer"
        },
        "page_size": {
          "type": "integer"
        },
        "total_rows": {
          "type": "integer"
        }
      }
    },
    "errcode.Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "msg": {
          "type": "string"
        }
      }
    },
    "model.Article": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "conver_image_url": {
          "type": "string"
        },
        "created_by": {
          "type": "string"
        },
        "created_on": {
          "type": "integer"
        },
        "desc": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "is_del": {
          "type": "integer"
        },
        "modified_by": {
          "type": "string"
        },
        "modified_on": {
          "type": "integer"
        },
        "state": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "model.ArticleSwagger": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/model.Article"
          }
        },
        "pager": {
          "type": "object",
          "$ref": "#/definitions/app.Pager"
        }
      }
    },
    "model.Tag": {
      "type": "object",
      "properties": {
        "created_by": {
          "type": "string"
        },
        "created_on": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "is_del": {
          "type": "integer"
        },
        "modified_by": {
          "type": "string"
        },
        "modified_on": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "state": {
          "type": "integer"
        }
      }
    },
    "model.TagSwagger": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/model.Tag"
          }
        },
        "pager": {
          "type": "object",
          "$ref": "#/definitions/app.Pager"
        }
      }
    }
  }
}